MA-18631 allow "fastboot reboot" command failure in flash script
After host issues "fastboot reboot" command, it then read the response
from the device. The device receives this command, send response and then
reset.

Very rarely, when the host try to read the response, the board already
reset and host cannot see the device, the response cannot be read, cause
failure of this command.

The flash script with "-e" option will exit when a command exit non-zero
value. This commit remove this limitation on "fastboot reboot" command.

Change-Id: I4e66648191571693c02d2d2ec15032e7ea41ec49
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/14913
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>