MA-17035 [Android_8MN_EVK] CTS Verifier: Camera FOV calibration test fail on all resolutions.
The issue is introduced by commit 4957249da4. The commit is to fix
testHwSupportedLevel failed due to "Active array size or
cropped active array size and max JPEG size should be similar".

In previously version, no 2592x1944 on evk_8mn. So change mActiveArrayWidth,
mActiveArrayHeight, PhysicalWidth and PhysicalHeight to align with 1080p.
But not scale the FocalLength with 1080/2592.

Now 2592x1944 is suppored, so we recover these configs to align with 2592x1944,
just as other boards.

Change-Id: I6f6ae72a01a45eb44a5d211bacf4da991e6e76ed
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15403
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16190
Reviewed-by: Tian Yang <yang.tian@nxp.com>