MA-18902 Fix screen blur issue when operate UI sometimes on 8mq board
DRM display fence will be signaled when the buffer start to display
on the screen instead of finish display, there is risk of the buffer
be overwrite by SurfaceFlinger composition when only 3 framebuffer.
For indirect rendering GPU will allocate one internal temp buffer
so totally have four buffers, but on 8qm/8qxp and 8mq with direct
rendering enable GPU render to framebuffer directly so framebuffer
count need enlarge to 4.

Change-Id: I413acccb8da62be524f1b60202d5b670e14092e3
Signed-off-by: Richard Liu <xuegang.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15609
Reviewed-by: juan.hao <juan.hao@nxp.com>