MA-18429-12 Note down the files need name modification in advance
Note down the files whose names to be modified in advance, then copy
files and check whether those file names are used in the file to be
copied.

os.listdir() list files in arbitrary order, At the time when this script
was developed and verified, althouth noting down files need name
modification and copy files are in the same loop, the "os.listdir()"
listed the files need name modification first and files which use these
names then, so the issue is not triggered.

Change-Id: I01c7487612439cb27e665391906099aa4d91d628
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15709
Reviewed-by: Wang Haoran <elven.wang@nxp.com>