MA-19095-2 set usb controller sysfs path for gadget HAL to watch on
Set value and security context for property "vendor.usb.sysfs_path".
The value is the sysfs path of usb controller, usb gadget HAL will
watch on this path based on this property.

Allow usb gadget HAL to add inotify watch on the path.

For platforms without this property set, USB hal will fail to add watch
on the "", then the HAL won't try to bind configfs with dummy udc.

Change-Id: I06c005301b4603530a54f30a2169c5c7802fe531
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16022
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>