MA-19215 support to specify the board to be flash by uuu script
Add the "-usb" option for uuu_imx_android_flash files, the parameter
followed by this option is a usb path. this option can be used multiple
time to specify more than one usb path. The boards connected to the
specified usb port will be flashed, so different boards can be flashed
concurrently.

The usb path can be found when use uuu to flash the image, it is at the
beginning of the line with progress bar like below, the board now being
flashed is connected to a port with path 2:2.

2:2     13/20 [==========>      31%                  ] FB[-t 600000]: flash super super.img

Change-Id: Ib780e6460b85d330f1c14d86939799988d2b4f68
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16259
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>