MA-19095-6 set udc name directly without using the prop set by init
In init code, property "sys.usb.controller" is set based on reading
entries under "/sys/classs/udc", the first entry whose name is not
started with "." is set to it. When there are multiple udc devices, this
value is not fixed.

Set the property "sys.usb.controller" and "vendor.usb.config" in .rc
file directly with the udc name, make the udc device to be used fixed.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I2ec096183b5560234198a43bd5478d67f490f44e
Reviewed-on: http://androidsource.ap.freescale.net/project/16298
Reviewed-by: Wang Haoran <elven.wang@nxp.com>