MA-19069 refine selinux workaround when bring up android S
The corresponding workaround patch has:
  commitID: 7c6b9a4031965e3d6a18ad0369d7200c7a75eb4c
  changeID: I0674dcffc1aeecf3296e1c3f649ff67a2ae4049e

The rule of allowing recovery to access ffs property was added to allow
minadbd to access "sys.usb.ffs.ready". Now a similar rule can be found
in "system/sepolicy". So this rule can be removed.

The rule of allowing vendor_init to access default_prop was added
because of property access in .rc files in vendor partition. Now
properties accessed in .rc files in vendor partition are not with
default_prop label, so this rule can be removed.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I572b93c3dafdf7dadf5deee6a1959705454da5c2
Reviewed-on: http://androidsource.ap.freescale.net/project/16576
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>