MA-19300 refactor the part of storage configurations
If try to do an OTA from old image set without system_ext partition,
change the value of BOARD_USES_SYSTEM_EXTIMAGE to be false does not
work, the system_ext.img is still built.

In this patch, system_ext partition related configs take effect only
when the value of BOARD_USES_SYSTEM_EXTIMAGE is true.

Also change the partition list of OTA and dynamic partitions based on
the features.

Change-Id: I9fb2561409540f7019a6e46ba6c24a55e6ff5e48
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16925
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: Wang Haoran <elven.wang@nxp.com>