MA-19498 utilize the virtual A/B with compression feature
Inherrit the "compression_with_xor.mk", then below properties are
set for releasetools to be checked when generate OTA package.

    ro.virtual_ab.compression.xor.enabled
    ro.virtual_ab.compression.enabled

This "compression_wit_xor.mk" is for devices with virtual A/B and
vendor_boot partition, it covers the virtual_ab_ota.mk.

After the file is inherrited, snapuserd will be integrated with
below modules. snapuserd need to be used in init first stage

    snapuserd.vendor_ramdisk
    snapuserd
    snapuserd.recovery

snapuserd need to be used in first stage ramdisk, set below config so
with "snapuserd.vendor_ramdisk", there will be a copy of snapuserd in
vendor_ramdisk/first_stage_ramdisk/system/bin/ insteand of
vendor_ramdisk/system/bin/ and it can be used in init first stage.

    BOARD_MOVE_RECOVERY_RESOURCES_TO_VENDOR_BOOT

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: Ic3a94d9aed537b3b784afc6045ae618d27557739
Reviewed-on: http://androidsource.ap.freescale.net/project/17073
Reviewed-by: ji.luo <ji.luo@nxp.com>