MA-19532 Fix Android12 SurfaceFlinger crash issue on 8mm
There is a crash issue from SurfaceFlinger when use SkiaGL RenderEngine,
the reason is F16 texture is not supported on 8mm GPU. This patch set
RenderEngine to gles to avoid the crash, it can be regarded as a formal
fix. But we still need further check if any way to avoid SkiaGL crash
by enhance the error handling in SurfaceFlinger and GPU driver.

Change-Id: I6fa66f38d3d02b61962481e3891937df2d200b5a
Signed-off-by: Richard Liu <xuegang.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17165
Reviewed-by: zhang bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17197
Reviewed-by: Wang Haoran <elven.wang@nxp.com>