MA-19554-2 build kernel/modules/dts with individual target
Define two new phony targets: KERNEL_MODULES and KERNEL_DTB, they are
used for building kernel in-tree modules and dts files.

Also, in kernel.mk, the target $(KERNEL_OUT) is moved forword to be the
first target, because with imx-make.sh, the make command may be executed
without a target specified.

In imx-make.sh, changes are made to:
  * when "kernel" is specified, only build the kernel image
  * when "vendorbootimage" and "vendorimage" are specified, kernel
    in-tree and out-of-tree modules are built because they contain
    kernel module files.

    vendorimage depends on ${OUT}/dtb.img, so when "vendorbootimage" is
    specified, kernel dts files are also built.
  * when "dtboimage" is specified, build the kernel dts files.

Add a new shell env named "skip_config_or_clean", it is used in
kernel.mk because the config and clean work just need to be done once in
the one time execution of imx-make.sh.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: Ib2f8b77d71aeb89ea46c6babffcf51c468c0c267
Reviewed-on: http://androidsource.ap.freescale.net/project/17245
Reviewed-by: Wang Haoran <elven.wang@nxp.com>