MA-19095-5 allow usb HAL to set two notification related props
Below two properties are used in the framework to decide whether adb
connection and charging status notifications should be updated.

  persist.adb.notify
  persist.charging.notify

Now these two properties are set in usb HAL to make notifications not
update when the configfs is bound to dummy UDC. Security context should
be first set for the two props.

With BOARD_SEPOLICY_DIRS/BOARD_VENDOR_SEPOLICY_DIRS to specify
the sepolicy directories, the property name has limitations on
Android12, so SYSTEM_EXT_PRIVATE_SEPOLICY_DIRS is used.

This is used for SoC with cdns3 IP.

Change-Id: I14798cc736609164cf7e4dee7495221bc22fc510
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17706
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>