MA-19095-6 set the security context of system props in system_ext
Below two properties are used in the system_server to decide whether
the notifications related to adb connection and charging status
should be updated.

  persist.adb.notify
  persist.charging.notify

These two properties are not labeld with security context, so their
security context is "default_prop". they are set in the imx USB HAL,
but imx uSB HAL are not allowed to access the props with "default_prop"
context.

These two properties should then be labled with a seurity context, and
related rules should be added. Setting property security context under
vendor partition has namespace limitation, that is to say the property
names should have some fixed prefixes.There are vts cases check on this.

In this commit, the security context of these two properties are set in
system_ext partition.

Change-Id: Ib4779e8686aeff8c880dc69809b53e0614bfb8d0
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/17714
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>