MA-19095-7 allow usb HAL to set two notification related props
allow usb HAL to set below two properties.

      persist.adb.notify
      persist.charging.notify

The selinux rules are previously not set for imx8m and imx8ulp based
devices because the gadget HAL of these devices does not bind configfs
with dummy udc, it's okay to deny the access. This time, to eliminate
the denial messages, set these rules.

Change-Id: Ie8ccb34f14be1c32711cb9bd5a66b455aba96d4d
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/18037
Reviewed-by: Wang Haoran <elven.wang@nxp.com>