MA-19325-5 make sign GKI boot.img to be the default behaviour
With the value of IMX_SIGN_GKI default not to be "true", when run
commands like "make otapackage", "make dist", the imx boot image is
integrated to the target package, not the GKI boot image.

To make the GKI boot image used by default when execute make command
without specifying any env var:

  * rename IMX_SIGN_GKI to be TARGET_IMX_KERNEL, it contols whether the
    imx kernel is used or the GKI kernel is used.
  * set the TARGET_IMX_KERNEL default value to be "false". so when execute
    make command without specify any env var, GKI boot image is used.

Now the only place needs to overwrite the default value of TARGET_IMX_KERNEL
is in the imx-make.sh.

Change-Id: Ica64754cd44bec0a23595292e751156a63d1ac70
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/18131
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>