MA-20133-1 bump the shipping api level to 31
After the PRODUCT_SHIPPING_API_LEVEL is set to be 31, if
PRODUCT_SET_DEBUGFS_RESTRICTIONS is not assigned a value in
device makefile, it is set to true by default in core makefile.

Set the PRODUCT_SET_DEBUGFS_RESTRICTIONS to be true in user build
images and false in userdebug/eng build images. then:

 * with userdebug/eng build images, dumpstate can work without
   triggering selinux denial. this may help in debugging.
 * with user build images, the debugfs access is limited.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: Id7d0b6dcf4422ac84981e9317bc0016dda256407
Reviewed-on: http://androidsource.ap.freescale.net/project/18425
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>