MA-20149 workaround to bring up android-t-preview-2 on imx devices
the clang version is changed based on the value of "ClangDefaultVersion"
under build/soong.

some selinux rules being made as comments because of violation of
neverallow rules.

the gsi_keys.mk is removed from the code.

trusty is disabled.

encryption options on userdata partition are all removed.

kernel target-level is changed based on the build error log.

MultiCamera is not integrated because of build error log.

Change-Id: I2bc030047751c9fb45928f309ad03f56dc30f32a
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>