MA-20274 use NCM when click on "USB tethering" instead of RNDIS
The RNDIS config item is removed from the gki_defconfig file with below
patch:

    commit 0ea619fbd3aa56e764ee12ca680423ff2b7ce386
    Author: Greg Kroah-Hartman <gregkh@google.com>
    Date:   Tue Mar 8 11:42:43 2022 +0100

        ANDROID: GKI: disable rndis USB gadget driver

In the commit message, it is said that "NCM is a much better interface
to use and it also works on other operating system hosts".

There is a config item which can be used to select the usb function when
click on the "USB Tethering" radio button. overlay this config item to
use NCM as default.

Although the defconfig files used for mek_8q platforms still have the
RNDIS configs, the overlay is also made.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I603698410ebd80a9b1511c75d27e8cf07a3c7e66
Reviewed-on: http://androidsource.ap.freescale.net/project/19018
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
(cherry picked from commit 47850b0fc93821d0ac6de8e0e80d2061aa082d53)
Reviewed-on: http://androidsource.ap.freescale.net/project/20090