MA-20355 Crash when take 720p/1080p picture under single os08a20 4k mode
Crash due to jpeg size(1913555) is larger then the dest buffer size(1769478)
"I CameraDeviceSessionHwlImpl: processJpegBuffer, dstbuf size 1769478, 4177920, jpeg_size 1913555, max jpeg size 6291456".

As in 1080p mode, the "MaxJpegSize" is 4MB, so the dstbuf size for 1080p picture is 4MB.
In 4k Mode, if "MaxJpegSize" is 6MB, the dstbuf size for 1080p is (6M*1920*1080)/(3840*2160), plus some margin, it's 1769478.
As theory, it should be 16MB. After test, 12MB is enough.

Change-Id: I14216ddb699d3bbd75280b8cca46c1a2bdd281ed
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19158
Reviewed-by: <anle.pan@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19312