MA-20618 flash the two slots if -a/-b is not specified
it is found that when use the imx-sdcard-partition.sh to flash images
with dual-bootloader enabled and without the use of -a/-b options, the
search of the target partition may get wrong result, an example is that
the boot.img will be flashed to bootloader_a partition.

in this commit, flash both two slots if -a/-b is not used.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I36f7256220aadaa856370633d9746ba3d5865428
Reviewed-on: http://androidsource.ap.freescale.net/project/19833
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>