MA-20667-2 enlarge the size of metadata partition for f2fs
Related test case is listed below.

  module: vts_virtual_ab_test
  case: Metadata#IsExt4

This case checks the value of property ro.vendor.api_level, if it is
less than 33, then the metadata partition should be formatted as ext4
or it should be formatted as f2fs.

The value of that property is related to below properties, with our
configuration, it is 33.

  {"ro.board.api_level", "ro.board.first_api_level"}
  {"ro.product.first_api_level", "ro.build.version.sdk"}

16MB and 32MB are not sufficient for f2fs volume, so 64MB is used.

Change-Id: I4e1a7af39a4c1f5f51e3d829aceb9df546077c24
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19933
Reviewed-by: Wang Haoran <elven.wang@nxp.com>
Reviewed-by: ji.luo <ji.luo@nxp.com>