MA-20667-3 mount metadata with f2fs options
Related test case is listed below.

  module: vts_virtual_ab_test
  case: Metadata#IsExt4

This case needs the metadata partition be formatted as f2fs with the
value of property ro.vendor.api_level no less than 33. So mount the
partition with f2fs options.

Also, when use the script to flash sdcard, format metadata partition as
f2fs if possible.

Change-Id: Id2332cdcf9628823566009ec0f836dfe8ee09454
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19934
Reviewed-by: ji.luo <ji.luo@nxp.com>