MA-20683 charger function support with the health aidl version
The health AIDL HAL service provides functionalities of charger. As a result,
the system charger at /system/bin/charger is deprecated, we need to remove it,
and the ro.charger.no_ui is not needed now.

check the health aidl implementation, there are 2 points need to highlight.
1. getDiskStats need to be refined, will implement this later.
2. charger picture is not loaded right, will push another patch to fix it.

except above 2 points, there is no any other thing need to modify on the default aidl hal.

Change-Id: I19a2ca9272086ae806badf74ba4046f4424caee0
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/19958
Reviewed-by: faqiang.zhu <faqiang.zhu@nxp.com>