MA-20879-1 Revert "allow usb HAL to set two notification related props"
This reverts two commits:
  574d94c7c638f5d300a7d7d7a831912f00a68780
  36cfc40927365873a834e7f7a3d6e83a69b4f651

The reverted commits are used to allow usb hal to access below two
property:

    persist.adb.notify
    persist.charging.notify

The changes made in the two commits are manually removed as the configfs
is not bind to dummy UDC now.

A file for watch_8ulp is also modified as it is copied from other file.

Change-Id: I7f090ffbc43b01dabd66d4a7debc8e811746305d
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20207
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>