MA-20879-2 Revert usb sysfs_path related changes
This commit removed the content related to below property:

    vendor.usb.sysfs_path

Its value is the sysfs path of the usb controller. the usb HAL watches
on the path and check whether the udc exists or not when there are
events in the path, if the udc disappears, then try to bind configfs
with dummy udc.

Now there is no need to bind to dummy udc, instead, another property is
needed to be set. The usb controller path have too many events even the
board is connected to the host without any change to the connection, so
it's not a good choice to use it now.

Change-Id: Iaaf05190df6aed24c0787ab4518d8e0652d66c39
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20208
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>