MA-20916 remove the temporary files when flash the images
When execute the script with sudo then without sudo, the second command
would always fail because the temporary of the same name cannot be
overwritten.

In this commit, append a random number to the end of the temporary
files, and remove the temporary files after the images are flashed or
the script is terminated.

Change-Id: If7891fe45011ea3cbb264c4c4753cf6c60e8ee73
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20260
Reviewed-by: Wang Haoran <elven.wang@nxp.com>