AAUTO-750 make bluetooth product.prop specific to non-Car builds
As the core/sysprop.mk contains wildcard, looking for product.prop file in $TARGET_DEVICE_DIR,
it includes the product.prop file even if we disable it in makefile. We need custom name for the
file thus the wildcard is not used by default.

This allows different bluetooth configuration for Car builds.

Signed-off-by: Ondrej Lutera <ondrej.lutera@nxp.com>
Change-Id: Iff4105b53f0748bdbbc894f3aacb5999faa5b688
Reviewed-on: http://androidsource.ap.freescale.net/project/20514
Reviewed-by: Jiri Rezler <jiri.rezler@nxp.com>