MA-21019 [uuu flash script] hand the path with spaces
If the path to the images to be flashed contains spaces, then when the
path is used in commands, it should be quoted or the space " " should be
replaced with "\ ", they cannot be used together.

This patch quotes the (sub)path which contains spaces.

When a path is specified when execute the script with "\ " to represent
a space, it is handled by shell and actually what the script get does
not contain that back slash, so there is no need to remove the "\" when
the path specified in the commands contains "\ ".

When a path started with "~" is specified when execute the script, it is
replaced with the value of $HOME before passing to the script, so remove
the judgement of whehter the patch is started with "~"

Change-Id: I8a2b60d01066a5c5692c49bc08835cd81251058c
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20976