MA-21154 fix VTS module VtsHalThermalV1_0TargetTest on imx8qxp
VtsHalThermalV1_0TargetTest failed on imx8qxp is caused for ThermalHAL could not
be initialized properly, this is due to could not find cpu-thermal0 in sysfs.

in the current kernel version, change cpu thermal zone name from "cpu-thermal0"
to "cpu0-thermal", so change thermal json file to align with the kernel sysfs.

Change-Id: I5bc47e587eea34a0789a8ab64d8049273da689ee
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/21066
Reviewed-by: zhang bo <bo.zhang@nxp.com>