MA-20618 flash the two slots if -a/-b is not specified
it is found that when use the imx-sdcard-partition.sh to flash images
with dual-bootloader enabled and without the use of -a/-b options, the
search of the target partition may get wrong result, an example is that
the boot.img will be flashed to bootloader_a partition.

in this commit, flash both two slots if -a/-b is not used.

Change-Id: Id78ee9bfd1ffdb5f800d39c9fea08bef62c8501b
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/21113
Reviewed-by: ji.luo <ji.luo@nxp.com>