MA-21214 remove the label for fuseblk
fuseblk is used to mount ntfs. Check the code, it can be known that imx6
and imx7 device configs include the ntfs related module of ntfs-3g and
ntfsfix, but these modules cannot be found in current android13 code.
Also tried to plug a udisk with NTFS to android device, it is not
recognized.

On android u beta code, in syste/sepolicy, there is a patch which also
labels the fuseblk filesystem. so the label operation in this repo need
to be removed. For the patch in system/sepolicy, when it is reviewed in
aosp, it is set with a topic, under this topic, there is another patch
that try to support ntfs in vold, but that patch is not landed.

Change-Id: Ib0297a4d7aac77654b287410c080e2fc266a4cff
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/21259
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>