MA-20916-2 do not append random number to temp files if dryrun is used
The -dryrun option is used to generate the uuu script which is directly
handled by uuu but not execute uuu.

Someone may need to use the generated uuu.lst file in other places.
append random numbers to the temp files causes the file names used in
the uuu.lst also have the random number, which is a bit unconvenient.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I1ff227a2888484e5dc26d69f50b3b27b3b02f3ac
Reviewed-on: http://androidsource.nxp.com/project/21629
Reviewed-by: Elven Wang <elven.wang@nxp.com>