MA-21258-3 add an optional parameter for the -m option in uuu script
For evk_8ulp, now there are two mcu image, one for lpa, and one for lpd.
their uboot images are the same, but now two copies of the uboot images
are built and the mcu image to flash is selected based on the parameter
for -u option.

Add an optional option for the -m option. In the future, it can be used
as below, no need to depend on the -u option parameter. then only one
copy of uboot image can be built.

    -m lpa
    -m lda

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I8b309b3fc493f2940a728095ca891a5f407af759
Reviewed-on: http://androidsource.nxp.com/project/21778
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
Reviewed-by: Elven Wang <elven.wang@nxp.com>