MA-21404 fix the syntax error in the batch script
for commands executed in a if block and not econnected with the &
symbol, put them in a parentheses, or, they are executed event if the
if condition is not right. and the shift command make some command
line parameters not correctly processed, it's the options handled after
-m in the script, for example, the -e parameter is not well handled
and cause the userdata partition not erased.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I5ab11408ab30c9504b9a3438b814e10c20e8f30a
Reviewed-on: http://androidsource.nxp.com/project/21922
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>
(cherry picked from commit 0ba6665015332affecf5594bf0ab79eaa9af4790)
Reviewed-on: http://androidsource.nxp.com/project/22865
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>