MA-21404-2 workaround for the condition of echo -e
The built-in command `echo` supports some options like -e, -E and -n.

`echo -e` outputs nothing.
`echo -a` outputs "-a".

The whether_in_array function in the shell script depends on the `echo`
command. As a workaround, add a "+" before the first parameter to echo
and then compared the result to the array element which also prefixed
with a "+".

This helps when -e is directly used after -m.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: If95e56b0de37b93e4976df2b24f6db5be40d0f80
Reviewed-on: http://androidsource.nxp.com/project/22018
Reviewed-by: Elven Wang <elven.wang@nxp.com>