MA-21690-2 build 64-bit-only image conditionally
beginning from shipping api level 34, the VTS requires the value of
property "ro.product.cpu.abilist32" to be empty.

inherit the core_64_bit_only.mk can make the value of that property to
be empty.

do not set value for TARGET_2ND* variables, 32bit libs and apps are not
integrated.

The enviroment varialbe named IMX8_BUILD_64BIT_ROOTFS or
IMX9_BUILD_64BIT_ROOTFS is used to control whether 64-bit-only images
are built. If it is properly set:
  * the api level is set to be 34
  * TARGET_2ND* variables are not set
  * core_64_bit_only.mk is inherited insteand of core_64_bit.mk

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: Ia205f269709d130239a5e3ba23ed0f05c24e5c13
Reviewed-on: http://androidsource.nxp.com/project/22794
Reviewed-by: Elven Wang <elven.wang@nxp.com>
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>