MA-21785-2 imx93: fix build break in imx-mkimage
Below commit separates the i.MX91/i.MX93/i.MX95 from i.MX9
folder:
  commit 6348e9c6de13ad112de12463ccf22d67bbdb14a0
  Author: Ye Li ye.li@nxp.com
  Date:   Fri Oct 27 11:34:51 2023 +0800

    iMX9: Separate iMX93/iMX91/iMX95 to different folders

    To support multiple platforms development simultaneously, separate
    iMX93/91/95 to different folders. In soc.mak, only keep valid build
    targets and settings for current platform.

  Signed-off-by: Ji Luo <ji.luo@nxp.com>

However, above commit only exists on the "master" branch but not
on the "lf-6.1.55_2.2.0" branch. To make the build pass on both
branches, this commit checks the folder in imx-mkimage and decides
the right path we should use.

Change-Id: I7baba76d10588c0f882f86d756096301cd895b01
Reviewed-on: http://androidsource.nxp.com/project/23032
Reviewed-by: Jindong Yue <jindong.yue@nxp.com>